[OT-178] [FIX]: Category 관련 API 분리
