+++ src/entities/auth/api/auth.ts
@@ -1,24 +1,5 @@
-import api from "@shared/api/apiClient"; 
-import { ApiResponse } from "@/shared/types";
-
 const BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
-
-export interface CategoryItem {
-  categoryId: number;
-  name: string;
-}
-
-export interface TagItem {
-  tagId: number;
-  name: string;
-}
 
 export const authApi = {
   getKakaoOAuthUrl: () => `${BASE_URL}oauth2/authorization/kakao`,
-  getCategories: () =>
-    api.get("/categories") as unknown as Promise<ApiResponse<CategoryItem[]>>,
-  getTags: (categoryId: number) =>
-    api.get(`/categories/${categoryId}/tags`) as unknown as Promise<ApiResponse<TagItem[]>>,
-  setPreferredTags: (tagsId: number[]) =>
-    api.post("/member/me/tags", { tagsId }),
 };
