+++ src/entities/auth/api/index.ts
@@ -1,3 +1,8 @@
 export { authApi } from "./auth";
-export type { CategoryItem } from "./auth";
-export  type { TagItem } from "./auth";
+export type { CategoryItem } from "./category";
+export type { TagItem } from "./category";
+export type { SetPreferredTagsRequest } from "./category";
+export type { SetPreferredTagsResponse } from "./category";
+export { getCategories } from "./category";
+export { getTags } from "./category";
+export { setPreferredTags } from "./category";
